Enables diverging a `Serve` trait's generic param to allow it's specialization for arbitrary types
Previously the `Serve` trait had two generic parameters, one for argument deserialization and another for
result serialization. But when third-party libraries tried to implement this trait for generic state types, since all of the
trait and it's generic parameters were from a third-party (this) crate, this implementation was not allowed by the compiler.

So a technique of using an extra generic argument slot to allow the user to "pin-down" the trait with a locally-defined type is now added
(and the required setup for this is done automatically by the macro).

See See [RFC 1023](https://github.com/rust-lang/rfcs/blob/master/text/1023-rebalancing-coherence.md) for more details.